% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CauseClassification.R
\name{CauseClassification}
\alias{CauseClassification}
\title{Identify cause of death by comparing with plp result files.}
\arguments{
\item{outputFolder}{}

\item{seedNum}{}
}
\description{
Identify cause of death by comparing with plp result files.
}
