% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausePrediction.R
\name{CausePrediction}
\alias{CausePrediction}
\title{Predict cause of death by comparing with plp prediction values in result files.}
\arguments{
\item{outputFolder}{your output folder}

\item{TAR}{Time At Risk window end}

\item{nTree}{Number of Tree of Random Forest model}

\item{seedNum}{Seed number}
}
\description{
Predict cause of death by comparing with plp prediction values in result files.
}
